+++ src/storage-preview/azext_storage_preview/vendored_sdks/azure_mgmt_preview_storage/v2020_08_01_preview/_storage_management_client.py
@@ -24,6 +24,7 @@ from .operations import StorageAccountsOperations
 from .operations import DeletedAccountsOperations
 from .operations import UsagesOperations
 from .operations import ManagementPoliciesOperations
+from .operations import BlobInventoryPoliciesOperations
 from .operations import PrivateEndpointConnectionsOperations
 from .operations import PrivateLinkResourcesOperations
 from .operations import ObjectReplicationPoliciesOperations
@@ -54,6 +55,8 @@ class StorageManagementClient(object):
     :vartype usages: azure.mgmt.storage.v2020_08_01_preview.operations.UsagesOperations
     :ivar management_policies: ManagementPoliciesOperations operations
     :vartype management_policies: azure.mgmt.storage.v2020_08_01_preview.operations.ManagementPoliciesOperations
+    :ivar blob_inventory_policies: BlobInventoryPoliciesOperations operations
+    :vartype blob_inventory_policies: azure.mgmt.storage.v2020_08_01_preview.operations.BlobInventoryPoliciesOperations
     :ivar private_endpoint_connections: PrivateEndpointConnectionsOperations operations
     :vartype private_endpoint_connections: azure.mgmt.storage.v2020_08_01_preview.operations.PrivateEndpointConnectionsOperations
     :ivar private_link_resources: PrivateLinkResourcesOperations operations
@@ -114,6 +117,8 @@ class StorageManagementClient(object):
         self.usages = UsagesOperations(
             self._client, self._config, self._serialize, self._deserialize)
         self.management_policies = ManagementPoliciesOperations(
+            self._client, self._config, self._serialize, self._deserialize)
+        self.blob_inventory_policies = BlobInventoryPoliciesOperations(
             self._client, self._config, self._serialize, self._deserialize)
         self.private_endpoint_connections = PrivateEndpointConnectionsOperations(
             self._client, self._config, self._serialize, self._deserialize)