+++ src/healthcareapis/HISTORY.rst
@@ -3,6 +3,19 @@
 Release History
 ===============
 
+0.3.1
+++++++
+
+* Add support for AcrConfiguration
+* Upgraded API version to latest stable
+* Fixed bugs
+
+0.3.0
+++++++
+
+* Add support for private link
+* Add support for customer managed key
+
 0.2.0
 +++++